fix(BJS2-40530): event organization test
-fix config of bean by name eventFilters, deleted the setter
-fix the controller method validation exception handler more correctly, created the class ControllerMethodArgumentValidationExceptionHandle
