+++ src/main/java/school/faang/user_service/exception/ControllerMethodArgumentValidationExceptionHandler.java
@@ -1,0 +1,33 @@
+package school.faang.user_service.exception;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Slf4j
+@ControllerAdvice
+public class ControllerMethodArgumentValidationExceptionHandler {
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<List<DataValidationException>> handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
+        List<DataValidationException> errors = new ArrayList<>();
+
+        e.getBindingResult().getAllErrors().forEach((error) -> {
+            String fieldName;
+            try {
+                fieldName = ((FieldError) error).getField();
+            } catch (ClassCastException ex) {
+                fieldName = error.getObjectName();
+            }
+            String message = error.getDefaultMessage();
+            errors.add(new DataValidationException(fieldName + " : " + message));
+        });
+        log.info(errors.toString());
+        return ResponseEntity.badRequest().body(errors);
+    }
+}
